Trim whitespace before and after url in CI
The URL for the source code for mason packages might have a space before it and this caused errors like "fatal: protocol ' https' is not supported" because of the leading space when trying to clone the repo in our CI checks.
This adds a simple step to trim the whitespace which will hopefully fix this failure.

Signed-off-by: Shreyas Khandekar <60454060+ShreyasKhandekar@users.noreply.github.com>